+++ stagemonitor-web/src/main/java/org/stagemonitor/collector/web/monitor/filter/HttpExecutionContextMonitorFilter.java
@@ -6,12 +6,7 @@ import org.stagemonitor.collector.core.StageMonitor;
 import org.stagemonitor.collector.core.monitor.ExecutionContextMonitor;
 import org.stagemonitor.collector.web.monitor.MonitoredHttpExecution;
 
-import javax.servlet.Filter;
-import javax.servlet.FilterChain;
-import javax.servlet.FilterConfig;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
+import javax.servlet.*;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
@@ -58,9 +53,11 @@ public class HttpExecutionContextMonitorFilter extends AbstractExclusionFilter i
 	protected void handleException(Exception e) throws IOException, ServletException  {
 		if (e instanceof IOException) {
 			throw (IOException) e;
-		} if (e instanceof ServletException) {
+		}
+		if (e instanceof ServletException) {
 			throw (ServletException) e;
-		} if (e instanceof RuntimeException) {
+		}
+		if (e instanceof RuntimeException) {
 			throw (RuntimeException) e;
 		} else {
 			throw new RuntimeException(e);